# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: SOC OS - SaaS Security Platform API (V30/V50)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List security events
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: tactic
          in: query
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of security events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
    post:
      operationId: processEvent
      tags: [events]
      summary: Process a security event through the full SOC pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Pipeline result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResult"
...
  /meta-cube/executions:
    get:
      operationId: listMetaCubeExecutions
      tags: [runtime]
      summary: List META-CUBE executions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, succeeded, failed, retrying, dead_letter, recovered]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: META-CUBE executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetaCubeExecution"
        "503":
          description: META-CUBE unavailable
    post:
      operationId: createMetaCubeExecution
      tags: [runtime]
      summary: Submit a META-CUBE execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetaCubeExecutionInput"
      responses:
        "202":
          description: Execution accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCubeExecution"
        "400":
          description: Invalid execution
        "503":
          description: META-CUBE unavailable
  /meta-cube/executions/{id}:
    get:
      operationId: getMetaCubeExecution
      tags: [runtime]
      summary: Get a META-CUBE execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: META-CUBE execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCubeExecution"
        "404":
          description: Execution not found
        "503":
          description: META-CUBE unavailable
  /meta-cube/executions/{id}/retry:
    post:
      operationId: retryMetaCubeExecution
      tags: [runtime]
      summary: Retry a failed META-CUBE execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCubeExecution"
        "404":
          description: Execution not found
        "409":
          description: Execution cannot be retried
        "503":
          description: META-CUBE unavailable
  /meta-cube/executions/{id}/recover:
    post:
      operationId: recoverMetaCubeExecution
      tags: [runtime]
      summary: Recover a META-CUBE execution from its checkpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Recovery accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCubeExecution"
        "404":
          description: Execution not found
        "409":
          description: Execution has no recoverable checkpoint
        "503":
          description: META-CUBE unavailable
  /meta-cube/dlq:
    get:
      operationId: listMetaCubeDlq
      tags: [runtime]
      summary: List META-CUBE dead-letter executions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Dead-letter executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetaCubeExecution"
        "503":
          description: META-CUBE unavailable
  /meta-cube/checkpoints:
    get:
      operationId: listMetaCubeCheckpoints
      tags: [runtime]
      summary: List META-CUBE checkpoints
      parameters:
        - name: executionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Runtime checkpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetaCubeCheckpoint"
        "503":
          description: META-CUBE unavailable

  # ─── TENANTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SecurityEvent:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        event:
          type: string
        score:
          type: integer
        action:
          type: string
          enum: [ALLOW, WARN, ISOLATE, PATCHED]
        status:
          type: string
          enum: [NEW, ACKNOWLEDGED, INVESTIGATING, QUARANTINED, RESOLVED]
        tactic:
          type: ["string", "null"]
        technique:
          type: ["string", "null"]
        techniqueId:
          type: ["string", "null"]
        velocityFlag:
          type: boolean
        correlationId:
          type: ["integer", "null"]
        nodeId:
          type: string
        hash:
          type: string
        prevHash:
          type: string
        timestamp:
          type: string
          format: date-time
        cpuUsage:
          type: ["integer", "null"]
        memUsage:
          type: ["integer", "null"]
      required:
        - id
        - event
        - score
        - action
        - status
        - velocityFlag
        - nodeId
        - hash
        - prevHash
        - timestamp
    EventInput:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        tenantId:
          type: ["integer", "null"]
        cpuUsage:
          type: ["integer", "null"]
        memUsage:
          type: ["integer", "null"]
        managedResourceKey:
          type: string
          pattern: "^[a-zA-Z0-9][a-zA-Z0-9._:-]{0,127}$"
        observedManagedState:
          type: object
          additionalProperties: true
      required: [event]
    PipelineResult:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/SecurityEvent"
        rulesMatched:
          type: integer
        scoreBoost:
          type: integer
        correlation:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/Correlation"
            - type: "null"
        pipelineStages:
          type: array
          items:
            type: string
        selfHealing:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/SelfHealingAction"
            - type: "null"
      required: [event, rulesMatched, scoreBoost, pipelineStages]
...
    SelfHealingAction:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        eventId:
          type: ["integer", "null"]
        restorePointId:
          type: ["integer", "null"]
        resourceKey:
          type: string
        location:
          type: string
        mode:
          type: string
          enum: [PREVIEW, APPLY, AUTO]
        status:
          type: string
          enum: [READY, RESTORED, NO_RESTORE_POINT, REJECTED, FAILED]
        previousStateHash:
          type: ["string", "null"]
        restoredStateHash:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, resourceKey, location, mode, status, message, createdAt]
...
    Correlation:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        threatType:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        confidence:
          type: number
        summary:
          type: string
        eventIds:
          type: string
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, threatType, severity, confidence, summary, eventIds, createdAt]
...
    MetaCubeExecutionInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        payload:
          type: object
          additionalProperties: true
        steps:
          type: array
          items:
            type: string
          minItems: 1
        idempotencyKey:
          type: string
        maxRetries:
          type: integer
          minimum: 0
          maximum: 10
      required: [name, payload, steps]
    MetaCubeExecution:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, retrying, dead_letter, recovered]
        payload:
          type: object
          additionalProperties: true
        steps:
          type: array
          items:
            type: string
        completedSteps:
          type: array
          items:
            type: string
        attempts:
          type: integer
        maxRetries:
          type: integer
        error:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, name, status, payload, steps, completedSteps, attempts, maxRetries, error, createdAt, updatedAt, finishedAt]
    MetaCubeCheckpoint:
      type: object
      properties:
        id:
          type: string
        executionId:
          type: string
        completedSteps:
          type: array
          items:
            type: string
        state:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required: [id, executionId, completedSteps, state, createdAt]